refactor: move Thruster, ChemicalThruster, ElectricThruster into thruster.py
